requirements.txt
python-telegram-bot==20.0
requests
pdfminer.six
python-dotenv
